ui/requirements.txt
fastapi
uvicorn[standard]
google-auth
google-cloud-storage
requests
